requirements.txt
et-xmlfile==1.1.0
json-tricks==3.17.2
Levenshtein==0.21.1
numpy==1.25.1
opencv-python==4.8.0.74
openpyxl==3.1.2
Pillow==10.0.0
pylibdmtx==0.1.10
